requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.4; python_version >= '3.6'
beautifulsoup4==4.9.3
django==3.2.4
djangorestframework==3.12.4
pytz==2021.1
soupsieve==2.2.1; python_version >= '3.0'
sqlparse==0.4.1; python_version >= '3.5'
